ID_bot/requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
idna==2.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
TelegramBotAPI==0.3.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.4

APScheduler==3.6.3
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
idna==2.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
TelegramBotAPI==0.3.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.4